almost done

+++ src/main/java/ru/yandex/practicum/filmrate/storage/FilmService.java
@@ -3,12 +3,11 @@ package ru.yandex.practicum.filmrate.storage;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.Collections;
 import java.util.Comparator;
 import java.util.List;
-import java.util.Set;
 import java.util.stream.Collectors;
 
+import ru.yandex.practicum.filmrate.exception.FilmIllegalArgumentException;
 import ru.yandex.practicum.filmrate.model.Film;
 
 @Service
@@ -23,25 +22,23 @@ public class FilmService {
     }
 
     public List<Film> getCountFilms(Integer count) {
-
         Comparator<Film> comparator = Comparator.comparingInt(film -> film.getLikeList().size());
-        List<Film> films = inMemoryFilmStorage.getAll();
-        Collections.sort(films, comparator);
-        List<Film> test = films.stream().limit(count).collect(Collectors.toList());
-        System.out.println("TEST " + test.toString());
-        return films.stream().limit(count).collect(Collectors.toList());
-    }
 
-    public Film getFilmById(Integer id) {
-        return inMemoryFilmStorage.getFilm(id);
+        return inMemoryFilmStorage.getAll().stream().sorted(comparator.reversed()).limit(count).collect(Collectors.toList());
     }
 
     public void deleteLike(Integer id, Integer userId) {
+        if (id < 0 || userId < 0) {
+            throw new FilmIllegalArgumentException("Айди не может быть отрицательным");
+        }
         inMemoryFilmStorage.getFilm(id).deleteLike(userId);
     }
 
     public void putLike(Integer id, Integer userId) {
-        System.out.println("Ставлю лайк");
+        if (id < 0 || userId < 0) {
+            throw new FilmIllegalArgumentException("Айди не может быть отрицательным");
+        }
+
         inMemoryFilmStorage.getFilm(id).addLike(userId);
     }
 }
